requirements.txt
flask
python-dotenv
psycopg2-binary
beeprint